CLN: move BaseMultiImageReader class to io.core
