+++ pyfibre/io/core/base_multi_image_reader.py
